% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abms.R
\name{rCRT}
\alias{rCRT}
\title{Title}
\usage{
rCRT(n, b, c)
}
\arguments{
\item{n}{Number of observations. It must be a positive integer.}

\item{b}{Parameter distribution, a non-negative integer. The number of Bernoulli independent variables that are added. It can be a vector}

\item{c}{Parameter distribution, a positive real number. Used calculate the success probability the j-th Bernoulli independent variable, that is, c/(c +j -1). It can be a vector}
}
\value{
This function generates \code{n} random variables from the \eqn{CRT}(\eqn{b,c}) distribution
}
\description{
Sampling from the Chines Restaurant distribution.
}
\examples{
#Generating 4 random variables with parameters b=2 and c=1
rCRT(4,2,1)
}
\references{
{Pitman, Jim (1995). "Exchangeable and Partially Exchangeable Random Partitions". Probability Theory and Related Fields. 102 (2): 145:158}
}
