% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abms.R
\name{summary_gibbs}
\alias{summary_gibbs}
\title{Summary function for abms objects}
\usage{
summary_gibbs(fit, BF = FALSE)
}
\arguments{
\item{fit}{An abms object. Such object is obtained by fitting a regression model with the \code{gibbs_abms()} function.}

\item{BF}{A logical object. if \code{TRUE}, then the Bayes factor comparison is shown.\code{BF=FALSE} by default.}
}
\value{
A summary of the inference performed by the Bayesian model obtained by the \code{gibbs_abms()} function. The variables are:

\item{Mean_IC}{A table with the posterior mean, standard deviation, and 95\% centered credible interval for each parameter}

\item{Explored_Models}{A table with all explored models. If \code{BF=TRUE}, the conditional Bayes factors and marginal Bayes factors estimator between the most probable model and the others that have arisen are displayed.}
}
\description{
For abms objects, it returns the posterior mean, standard deviation, and 95\% centered credible interval for each parameter. Additionally, it provides all explored models alongside the conditional Bayes factors and marginal Bayes factors estimator between the most probable model and the others that have arisen.
}
\examples{
## See \code{gibbs_abms()} help page function
}
