% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abms.R
\name{womack}
\alias{womack}
\title{For internal use
Womack prior}
\usage{
womack(K, rho)
}
\arguments{
\item{K}{Number of predictors. It must be a positive integer.}

\item{rho}{Value for the "rho" parameter. It must be positive real number}
}
\value{
Given that all models of the same hierarchy has the same prior probability, this function returns one value for each hierarchy.
including the null model (size=0)
}
\description{
Womack probability mass function with 'K' predictors and parameter 'rho'.
}
\examples{
#Fixing rho=1 and 3 predictors
womack(K=3, rho=1)
#it returns Womack prior for all models of size 0 (the null model), 1,2 and 3 (the full model)
}
\references{
{Womack, A., Fuentes, C., and Rodriguez-Taylor, D. (2015). "Model Space Priors for Objective Sparse Bayesian Regression." arXiv:1511.04745. 8:24}
}
